Move pjrt object build to pjrt dir
This starts separating out the pjrt plugin
from the jaxlib extension, since these
may need to be versioned separately in the future.

Either way, we'd like to pull the kernels from upstream
similarly to how we build PJRT from XLA sources, so
separating these sources out is the first step.