Fix devenv build on 0.6.0 (#144)
* Fix rocm_version substitution.

The string to be substituted changed in its comment section. This patch realigns
the search pattern with the new format and fixes the build failure caused by its
misalignment.

* Unpin python from version 3.10.

Ubuntu 24.04 does not have a package for python 3.10. Using the generic python3
series of packages will result in the distribution choosing the right one.

The patch also disables the interactive interface of apt so that the script
won't present menus to the user.

* Bump ROCm version from 6.4.0 to 7.0.2.

* Reformatting using black.

* Add missing doc strings.

* Set encoding to utf-8 in file open.

* Remove namespace patching.

The step is not required anymore.

* Update requirements to fetch jaxlib from 0.6.0 release.

Add requirements and python version guards to install the correct jaxlib from
the 0.6.0 release according to the python version used.