Avoid skips in nightly (#181)
* Enable unfiltered execution of the UTs during nightly builds.

The patch uses the `-s` (--ignore_skipfile) to re-enable the execution of the
tests in the UT suite.

* Remove the skip list file from master.

Master does not need to skip tests. We will reserve the feature occasionally for
specific branches.

* Remove the skip list from run_multi_gpu.py.

The run_multi_gpu.py script includes an API to skip specific tests. These are
not relevant for our master branch. This commit clears the implementation.