Refactor Docker image creation in two incremental steps.
The idea behind the refactoring is to provide an extra layer of indirection so
that we can reuse and pre-build the base containers in our CI. The advantage
will be two fold. First we can avoid rebuilding Python (and LLVM) in our CI for
every tasks by setting up a workflow that will produce the base images
periodically. Second, we can provide the same images but with LLVM as pre-canned
development environments.

NOTE: the Dockerfile should point to the GitHub registry. I'll fix the
references once we have ironed out the details.