Remove buggy device count patch (#227)
Putting a call to hipInit inside of jaxlib and calling it before we
load the PJRT plugin breaks profiling on ROCm. The rocprofiler
library must be initialized before hipInit is called, and the init
for rocprofiler hapens in PJRT. Removing this patchh will break
the wheel on systems that don't have ROCm installed, however. We
plan to fix that for the next release.